Cleaner proxy management in deployer
