+++ packages/core-modules/hardhat.config.js
@@ -1,6 +1,7 @@
 require('solidity-coverage');
 require('@nomiclabs/hardhat-ethers');
 require('@synthetixio/deployer');
+require('@synthetixio/cli');
 
 module.exports = {
   solidity: {
