+++ packages/deployer/internal/abis.js
@@ -1,0 +1,32 @@
+const UPGRADE_ABI = [
+  {
+    inputs: [],
+    name: 'getImplementation',
+    outputs: [
+      {
+        internalType: 'address',
+        name: '',
+        type: 'address',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'address',
+        name: 'newImplementation',
+        type: 'address',
+      },
+    ],
+    name: 'upgradeTo',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+];
+
+module.exports = {
+  UPGRADE_ABI,
+};
